% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ts.R
\name{compare_acf}
\alias{compare_acf}
\title{Calculate both Classical and Robust ACF}
\usage{
compare_acf(xt, ...)
}
\arguments{
\item{xt}{A data set}
}
\description{
Computes Classical & Robust ACF and packages it alongside the TS
}

