% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_fitted}
\alias{diag_fitted}
\title{Bind Fitted and Residual Values}
\usage{
diag_fitted(fitted, resid, std = FALSE)
}
\arguments{
\item{fitted}{A \code{numeric vector} containing fitted values from the model.}

\item{resid}{A \code{numeric vector} containing residuals.}

\item{std}{A \code{boolean} indicating whether to standardize the residuals}
}
\value{
A \code{data.frame} with structure
\itemize{
\item{fitted}{Fitted Values}
\item{residuals}{Residuals}
}
}
\description{
Creates a \code{data.frame} containing fitted and residual values from a
model.
}
\details{
This function will be reworked at a later time to provide model support.
}

