% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_qq}
\alias{diag_qq}
\alias{diag_qq.Arima}
\alias{diag_qq.default}
\alias{diag_qq.lm}
\alias{diag_qq.ts}
\title{QQ Normal plot functionality}
\usage{
diag_qq(x, std = FALSE)

\method{diag_qq}{Arima}(x, std = FALSE)

\method{diag_qq}{lm}(x, std = FALSE)

\method{diag_qq}{ts}(x, std = FALSE)

\method{diag_qq}{default}(x, std = FALSE)
}
\arguments{
\item{x}{A \code{\link{arima}}, \code{\link{lm}}, or a data object.}

\item{std}{A \code{boolean} indicating whether the residuals should be
standardized.}
}
\value{
A \code{data.frame} with structure
\itemize{
\item{residuals}{Residuals}
\item{quantiles}{Quantiles}
\item{mirror}{Line y-values}
}
}
\description{
Provides the calculations behind a qq normal plot.
}

