% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_resid}
\alias{diag_resid}
\alias{diag_resid.Arima}
\alias{diag_resid.default}
\alias{diag_resid.lm}
\title{Residuals of the Process}
\usage{
diag_resid(x, std = FALSE)

\method{diag_resid}{Arima}(x, std = FALSE)

\method{diag_resid}{lm}(x, std = FALSE)

\method{diag_resid}{default}(x, std = FALSE)
}
\arguments{
\item{x}{A \code{\link{arima}}, \code{\link{lm}}, or a data object.}

\item{std}{A \code{boolean} indicating whether the residuals should be
standardized.}
}
\description{
Extracts the residuals of the process
}

