% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_ts}
\alias{diag_ts}
\title{Time Series Diagnostics}
\usage{
diag_ts(model, xt, test = "ljung-box", robust = FALSE)
}
\arguments{
\item{model}{An \code{\link{arima}} object.}

\item{xt}{The data used to construct said model.}

\item{test}{A \code{string} indicating the type of portmanteau test to run.}

\item{robust}{A \code{boolean} indicating whether the ACF should be robust (\code{TRUE}) or
not (\code{FALSE}).}

\item{std}{A \code{boolean} indicating whether residuals should be standardized.}
}
\value{
A \code{diag_ts} object
}
\description{
Provides a list structure containing time series diagnostic information
}

