% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{diag_wv}
\alias{diag_wv}
\alias{diag_wv.Arima}
\alias{diag_wv.default}
\alias{diag_wv.diag_resid}
\alias{diag_wv.lm}
\title{Wavelet Variance (WV) fit for White Noise}
\usage{
diag_wv(x)

\method{diag_wv}{diag_resid}(x)

\method{diag_wv}{lm}(x)

\method{diag_wv}{Arima}(x)

\method{diag_wv}{default}(x)
}
\arguments{
\item{x}{A \code{\link{arima}}, \code{\link{lm}}, or data set object.}
}
\value{
A \code{diag_wv} object that inherits \code{gmwm} object information.
}
\description{
Computes whether the residuals match with a
White Noise by using the Generalized Method of Wavelet Moments (GMWM)
}

