% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ts.R
\name{eda_change}
\alias{eda_change}
\title{Viable Transforms}
\usage{
eda_change(x, t1 = diff, t2 = log, both = FALSE)
}
\arguments{
\item{x}{A data object}

\item{t1, t2}{A \code{function} to transform the data}

\item{both}{A \code{boolean} to apply both transforms to the data
e.g \code{t1(t2(x))}  (\code{TRUE}) or not e.g. \code{t2(x)} (\code{FALSE}).}
}
\value{
A \code{eda_change} object with
\itemize{
\item{original}{Untouched data}
\item{transform1}{First Transform}
\item{transform2}{Second Transform}
}
}
\description{
Provides the ability to graph multiple time series under different
conditions
}

