% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ts.R
\name{eda_ts}
\alias{eda_ts}
\title{EDA on Time Series}
\usage{
eda_ts(xt, lag.max = 30, robust = FALSE, ...)
}
\arguments{
\item{xt}{A data set}

\item{robust}{A \code{bool} indicating whether to robustly estimate only the
ACF (not PACF).}
}
\description{
Computes P/ACF and packages it alongside the TS
}

