% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{emp_acf}
\alias{emp_acf}
\alias{emp_acf.Arima}
\alias{emp_acf.default}
\alias{emp_acf.lm}
\title{Empirical ACF Plot}
\usage{
emp_acf(x, type = "correlation", robust = FALSE, ...)

\method{emp_acf}{Arima}(x, type = "correlation", robust = FALSE, ...)

\method{emp_acf}{lm}(x, type = "correlation", robust = FALSE, ...)

\method{emp_acf}{default}(x, type = "correlation", robust = FALSE, ...)
}
\arguments{
\item{x}{A data set, \code{\link{arima}} or \code{\link{lm}}}

\item{type}{A \code{string} with either \code{"correlation"} or \code{"covariance"}}

\item{robust}{A \code{boolean} indicating whether to use a robust estimation.}
}
\description{
Empirical ACF Plot
}

