% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{emp_pacf}
\alias{emp_pacf}
\alias{emp_pacf.Arima}
\alias{emp_pacf.default}
\alias{emp_pacf.lm}
\title{Empirical PACF}
\usage{
emp_pacf(x, ...)

\method{emp_pacf}{Arima}(x, ...)

\method{emp_pacf}{lm}(x, ...)

\method{emp_pacf}{default}(x, ...)
}
\arguments{
\item{x}{A data set, \code{\link{arima}} or \code{\link{lm}}}
}
\description{
Computes the empirical partial autocorrelation (PACF).
}

