% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{is.diag_resid}
\alias{is.diag_resid}
\title{Check is class}
\usage{
is.diag_resid(x)
}
\arguments{
\item{x}{A \code{diag_resid} object}
}
\value{
A \code{boolean} indicating \code{TRUE} or \code{FALSE}
}
\description{
Performs a check to see inheritance
}

