% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{plot.acf_plot}
\alias{autoplot.acf_plot}
\alias{plot.acf_plot}
\title{Plot ACF-like graphs}
\usage{
\method{plot}{acf_plot}(x, show.ci = TRUE, ci = 0.95, ...)

\method{autoplot}{acf_plot}(object, show.ci = TRUE, ci = 0.95, ...)
}
\arguments{
\item{x, object}{A object from \code{\link{emp_acf}}, \code{\link{emp_pacf}},
\code{\link{theo_acf_arma}}, or \code{\link{theo_pacf_arma}}.}

\item{show.ci}{A \code{bool} indicating whether to show confidence region}

\item{ci}{A \code{double} containing the 1-alpha level. Default is 0.95}

\item{...}{Additional parameters}
}
\description{
Generic ACF plot function for empirical and theoretical graphs
}
\examples{
# Calculate the Autocorrelation
m = emp_acf(datasets::AirPassengers)

# Plot with 95\% CI
plot(m)

# Plot with 90\% CI
plot(m, ci = 0.90)

# Plot without 95\% CI
plot(m, show.ci = FALSE)
}

