% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{plot.diag_fitted}
\alias{autoplot.diag_fitted}
\alias{plot.diag_fitted}
\title{Plot Fitted vs. Residual}
\usage{
\method{plot}{diag_fitted}(x, ...)

\method{autoplot}{diag_fitted}(object, ...)
}
\arguments{
\item{x, object}{A \code{\link{diag_fitted}} object}
}
\description{
Generates a fitted vs. residual graph
}

