% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{plot.diag_qq}
\alias{autoplot.diag_qq}
\alias{plot.diag_qq}
\title{QQ Normal Plot}
\usage{
\method{plot}{diag_qq}(x, ...)

\method{autoplot}{diag_qq}(object, ...)
}
\arguments{
\item{x, object}{A \code{\link{diag_qq}} object}

\item{...}{Additional parameters (not used)}
}
\description{
Generates a QQ Normal plot
}

