% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{plot.diag_resid}
\alias{autoplot.diag_resid}
\alias{plot.diag_resid}
\title{Graph the Distribution of Residuals}
\usage{
\method{plot}{diag_resid}(x, type = "hist", ...)

\method{autoplot}{diag_resid}(object, type = "hist", ...)
}
\arguments{
\item{x, object}{A \code{\link{diag_resid}} object.}

\item{type}{A \code{string} indicating either:
\code{"hist"} (residual histogram), \code{"resid"} (residual plot),
or \code{"both"}}
}
\value{
A \code{ggplot2} object.
}
\description{
Makes a histogram of the distribution of residuals
}
\seealso{
\code{\link[gmwm]{gmwm}} and \code{\link[gmwm]{plot.gmwm}}
}

