% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{plot.diag_ts}
\alias{autoplot.diag_ts}
\alias{plot.diag_ts}
\title{Graph Time Series Diagnostic}
\usage{
\method{plot}{diag_ts}(x, ...)

\method{autoplot}{diag_ts}(object, ...)
}
\arguments{
\item{x, object}{A \code{\link{diag_ts}} object.}
}
\value{
A \code{grid.arrange} object.
}
\description{
Creates a time series diagnostic plot
}

