% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_plots.R
\name{plot.diag_wv}
\alias{autoplot.diag_wv}
\alias{plot.diag_wv}
\title{Graph the Wavelet Variance (WV) Estimation of the White Noise}
\usage{
\method{plot}{diag_wv}(x, ...)

\method{autoplot}{diag_wv}(object, ...)
}
\arguments{
\item{x, object}{A \code{\link{diag_wv}} object.}
}
\value{
A \code{ggplot2} object.
}
\description{
Creates a plot containing the empirical WV and the implied WV for an
assumed white noise process.
}
\seealso{
\code{\link[gmwm]{gmwm}} and \code{\link[gmwm]{plot.gmwm}}
}

