% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ts.R
\name{plot.eda_change}
\alias{autoplot.eda_change}
\alias{plot.eda_change}
\title{Plot Different Time Series Transforms}
\usage{
\method{plot}{eda_change}(x, ...)

\method{autoplot}{eda_change}(object, ...)
}
\arguments{
\item{x, object}{A \code{\link{eda_change}} object}
}
\description{
Graphs time series transforms together with original data
}

