% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eda_ts.R
\name{plot.eda_ts}
\alias{autoplot.eda_ts}
\alias{plot.eda_ts}
\title{Visualize EDA on Time Series}
\usage{
\method{plot}{eda_ts}(x, ...)

\method{autoplot}{eda_ts}(object, ...)

\method{plot}{eda_ts}(x, ...)

\method{autoplot}{eda_ts}(object, ...)
}
\arguments{
\item{x, object}{A \code{eda_ts} object}

\item{x, object}{A \code{eda_ts} object}
}
\description{
Generates a time series EDA plot

Generates a time series EDA plot
}

