% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{plot.emp_corr}
\alias{autoplot.emp_corr}
\alias{plot.emp_corr}
\title{Plot Empirical Autocorrelation (ACF) for ARMA Models}
\usage{
\method{plot}{emp_corr}(x, ...)

\method{autoplot}{emp_corr}(object, ...)
}
\arguments{
\item{x, object}{An \code{\link{emp_corr}} using \code{\link{emp_acf}}
and  \code{\link{emp_pacf}}.}

\item{...}{Additional parameters}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Displays the empirical autocorrelation for ARMA Models
}

