% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_arima.R
\name{plot.select_arima}
\alias{autoplot.select_arima}
\alias{plot.select_arima}
\title{Plot ARIMA Selection Matrix}
\usage{
\method{plot}{select_arima}(x, ...)

\method{autoplot}{select_arima}(object, ...)
}
\arguments{
\item{x, object}{An object that is either of type \code{\link{select_arima}}
or \code{\link{select_arma}}.}
}
\description{
Constructs a facet wrapped graph showing model selection
criteria over multiple terms
}

