% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{plot.theo_arma}
\alias{autoplot.theo_arma}
\alias{plot.theo_arma}
\title{Plot Theoretical Autocorrelation (ACF) for ARMA Models}
\usage{
\method{plot}{theo_arma}(x, ...)

\method{autoplot}{theo_arma}(object, ...)
}
\arguments{
\item{x, object}{An \code{"theo_arma"} object from \code{\link{theo_acf}}
or  \code{\link{theo_pacf}}.}

\item{...}{Additional parameters}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Displays the theoretical autocorrelation for ARMA Models
}
\examples{
# Compute Theoretical ACF
m = theo_acf(ARMA(ar = -0.25, ma = NULL), lag.max = 7)

# Compute Theoretical PACF
m2 = theo_pacf(ARMA(ar = -0.25, ma = NULL), lag.max = 7)

# Plot either the theoretical ACF or PACF
plot(m); plot(m2)
}

