% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{plot.theo_corr}
\alias{autoplot.theo_corr}
\alias{plot.theo_corr}
\title{Plot Theoretical Autocorrelation (ACF) for ARMA Models}
\usage{
\method{plot}{theo_corr}(x, ...)

\method{autoplot}{theo_corr}(object, ...)
}
\arguments{
\item{x, object}{An \code{"theo_arma"} object from \code{\link{theo_acf_arma}}
or  \code{\link{theo_pacf_arma}}.}

\item{...}{Additional parameters}
}
\value{
An \code{array} of dimensions \eqn{N \times S \times S}{N x S x S}.
}
\description{
Displays the theoretical autocorrelation for ARMA Models
}

