% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_arima.R
\name{plot.select_comp}
\alias{autoplot.select_comp}
\alias{plot.select_comp}
\title{Visualize Select Model Functions}
\usage{
\method{plot}{select_comp}(x, ...)

\method{autoplot}{select_comp}(object, ...)
}
\arguments{
\item{x, object}{An object that is either of type \code{\link{select_ar}}
or \code{\link{select_ma}}.}
}
\description{
Provides a model comparison
}

