% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{theo_acf}
\alias{theo_acf}
\title{Computes the Theoretical Autocorrelation (ACF) of an ARMA process}
\usage{
theo_acf(model = ARMA(ar = c(0.5, -0.25), ma = 0.2), lag.max = NULL)
}
\arguments{
\item{lag.max}{A \code{integer} indicating the max length. Default is \eqn{max(ar, ma +1)}.}

\item{ar}{A \code{vector} containing the AR coefficients}

\item{ma}{A \code{vector} containing the MA coefficients}
}
\description{
Acts as a wrapper around `ARMAacf` to calculate the theoretical ACF process
of an ARMA process.
}
\examples{
# Computes the theoretical ACF for an ARMA(1,0) or better known as an AR(1)
theo_acf(ARMA(ar = -0.25, ma = NULL), lag.max = 7)
# Computes the theoretical ACF for an ARMA(2, 1)
theo_acf(ARMA(ar = c(.50, -0.25), ma = 0.20), lag.max = 10)
}

