% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{theo_corr}
\alias{theo_corr}
\title{Compute Theoretical ACF and PACF}
\usage{
theo_corr(model = ARMA(ar = c(0.5, -0.25), ma = 0.2), lag.max = NULL)
}
\arguments{
\item{lag.max}{A \code{integer} indicating the max length. Default is \eqn{max(ar, ma +1)}.}
}
\description{
Computes both the ACF and PACF for a given process.
}
\examples{
# Computes the theoretical ACF for an ARMA(2, 1)
theo_corr(ARMA(ar = c(.50, -0.25), ma = 0.20), lag.max = 10)
}

