% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.R
\name{emp_corr}
\alias{emp_corr}
\title{Compute Empirical ACF and PACF}
\usage{
emp_corr(x, robust = FALSE)
}
\arguments{
\item{x}{A data set, \code{\link{arima}} or \code{\link{lm}}}

\item{robust}{A \code{boolean} indicating whether to use a robust estimation.}
}
\description{
Computes both the ACF and PACF for a given dataset.
}
\examples{
# Computes the empirical correlation
emp_corr(gnp)
}

